<Update> Change SceneCamSorter to CameraManager
Removed singleton aspect with static functions
