+++ FlexEngine/FlexEngine.vcxproj.filters
@@ -214,7 +214,7 @@
     <ClCompile Include="src\FlexEngine\Renderer\OpenGL\opengltextrenderer.cpp">
       <Filter>src\FlexEngine\Renderer\OpenGL</Filter>
     </ClCompile>
-    <ClCompile Include="src\FlexEngine\Renderer\scenecamsorter.cpp">
+    <ClCompile Include="src\FlexEngine\Renderer\cameramanager.cpp">
       <Filter>src\FlexEngine\Renderer</Filter>
     </ClCompile>
     <ClCompile Include="src\FlexEngine\flexprofiler.cpp">
@@ -417,7 +417,7 @@
     <ClInclude Include="src\FlexEngine\Renderer\OpenGL\opengltextrenderer.h">
       <Filter>src\FlexEngine\Renderer\OpenGL</Filter>
     </ClInclude>
-    <ClInclude Include="src\FlexEngine\Renderer\scenecamsorter.h">
+    <ClInclude Include="src\FlexEngine\Renderer\cameramanager.h">
       <Filter>src\FlexEngine\Renderer</Filter>
     </ClInclude>
     <ClInclude Include="src\FlexEngine\flexprofiler.h">
