+++ FlexEngine/src/FlexEngine/Renderer/OpenGL/openglspriterenderer.h
@@ -30,7 +30,7 @@
 #include "flx_api.h"
 #include "FlexMath/vector4.h"
 #include "opengltexture.h"
-#include "../scenecamsorter.h"
+#include "../cameramanager.h"
 
 #include <glad/glad.h>
