+++ MicroChess/imgui.ini
@@ -9,8 +9,8 @@ Size=400,400
 Collapsed=0
 
 [Window][Statistics]
-Pos=996,37
-Size=284,468
+Pos=833,37
+Size=447,468
 Collapsed=0
 DockId=0x00000004,1
 
@@ -31,8 +31,8 @@ Collapsed=0
 DockId=0x00000001,0
 
 [Window][Inspector]
-Pos=996,37
-Size=284,468
+Pos=833,37
+Size=447,468
 Collapsed=0
 DockId=0x00000004,0
 
@@ -55,13 +55,13 @@ DockId=0x00000007,0
 
 [Window][Scene]
 Pos=252,37
-Size=742,468
+Size=579,468
 Collapsed=0
 DockId=0x00000002,0
 
 [Window][Game]
 Pos=252,37
-Size=742,468
+Size=579,468
 Collapsed=0
 DockId=0x00000002,1
 
@@ -76,11 +76,11 @@ Column 0  Weight=1.0000
 Column 1  Weight=1.0000
 
 [Docking][Data]
-DockSpace       ID=0x8B93E3BD Window=0xA787BDB4 Pos=640,382 Size=1280,713 Split=Y
+DockSpace       ID=0x8B93E3BD Window=0xA787BDB4 Pos=358,129 Size=1280,713 Split=Y
   DockNode      ID=0x00000005 Parent=0x8B93E3BD SizeRef=1280,468 Split=X
-    DockNode    ID=0x00000003 Parent=0x00000005 SizeRef=994,713 Split=X
+    DockNode    ID=0x00000003 Parent=0x00000005 SizeRef=831,713 Split=X
       DockNode  ID=0x00000001 Parent=0x00000003 SizeRef=250,713 Selected=0x2E9237F7
-      DockNode  ID=0x00000002 Parent=0x00000003 SizeRef=742,713 CentralNode=1 Selected=0xE192E354
-    DockNode    ID=0x00000004 Parent=0x00000005 SizeRef=284,713 Selected=0xE7039252
-  DockNode      ID=0x00000007 Parent=0x8B93E3BD SizeRef=1280,243 Selected=0x6F7CD859
+      DockNode  ID=0x00000002 Parent=0x00000003 SizeRef=579,713 CentralNode=1 Selected=0x26816F31
+    DockNode    ID=0x00000004 Parent=0x00000005 SizeRef=447,713 Selected=0xE7039252
+  DockNode      ID=0x00000007 Parent=0x8B93E3BD SizeRef=1280,243 Selected=0xD44407B5
